tidy router lookup and hash guards
