+++ lib/llm/src/kv_router.rs
@@ -18,7 +18,7 @@ use dynamo_kv_router::{
         RouterBackpressureReason, RouterEvent, RouterRequest, RouterResponse, RoutingConstraints,
         TokensWithHashes, WorkerConfigLike, WorkerId, WorkerWithDpRank, compute_block_hash_for_seq,
     },
-    scheduling::{OverloadedWorkerProvider, TierOverlapBlocks},
+    scheduling::OverloadedWorkerProvider,
 };
 use dynamo_runtime::{
     component::{Client, Endpoint},
@@ -33,7 +33,6 @@ use dynamo_runtime::{
     traits::DistributedRuntimeProvider,
 };
 use futures::stream;
-use serde::Serialize;
 use tracing::Instrument;
 use validator::Validate;
 
@@ -49,7 +48,9 @@ pub mod metrics;
 pub mod prefill_router;
 pub mod publisher;
 pub mod push_router;
+mod route_lookup;
 pub mod scheduler;
+mod scheduler_inputs;
 pub mod sequence;
 pub mod shared_cache;
 pub mod sticky_sessions;
@@ -58,15 +59,20 @@ pub use agent_controller::AgentController;
 pub use indexer::{Indexer, ServedIndexerHandle, ServedIndexerMode, ensure_served_indexer_service};
 pub use prefill_router::PrefillRouter;
 pub use push_router::{DirectRoutingRouter, KvPushRouter};
+pub use scheduler_inputs::{OverlapScoresResponse, SharedCacheOverlapScore, WorkerOverlapScore};
 pub use sticky_sessions::StickySessionRouter;
+
+use route_lookup::{TieredLookupResult, query_tiered_matches, split_retained_block_hashes};
+use scheduler_inputs::{
+    CacheHitEstimates, KvRouterOverlapRefresher, WorkerCacheHitEstimate,
+    cache_hit_estimates_from_tiered_matches, cache_hit_for_worker, shared_cache_overlap_score,
+    tier_overlap_blocks_from_tiered_matches,
+};
 
 use crate::{
     discovery::RuntimeConfigWatch,
     kv_router::{
-        scheduler::{
-            DefaultWorkerSelector, KvScheduler, OverlapScoresRefresh, PotentialLoad,
-            RefreshedOverlap,
-        },
+        scheduler::{DefaultWorkerSelector, KvScheduler, PotentialLoad},
         sequence::{SequenceError, SequenceRequest},
     },
     local_model::runtime_config::ModelRuntimeConfig,
@@ -85,24 +91,6 @@ pub enum FindBestMatchOutcome {
         queued_isl_tokens: usize,
         max_queued_isl_tokens: Option<usize>,
     },
-}
-
-fn split_retained_block_hashes(
-    block_hashes: Vec<LocalBlockHash>,
-    supports_overlap_refresh: bool,
-    return_routing_hashes: bool,
-) -> (Option<Vec<LocalBlockHash>>, Option<Vec<LocalBlockHash>>) {
-    debug_assert!(supports_overlap_refresh || return_routing_hashes);
-
-    if supports_overlap_refresh && return_routing_hashes {
-        return (Some(block_hashes.clone()), Some(block_hashes));
-    }
-    if supports_overlap_refresh {
-        return (Some(block_hashes), None);
-    }
-
-    debug_assert!(return_routing_hashes);
-    (None, Some(block_hashes))
 }
 
 // [gluo TODO] shouldn't need to be public
@@ -124,256 +112,6 @@ pub const RADIX_STATE_FILE: &str = "radix-state";
 
 // for worker-local kvindexer query
 pub const WORKER_KV_INDEXER_BUFFER_SIZE: usize = 1024; // store 1024 most recent events in worker buffer
-
-#[derive(Debug, Clone, Copy, Default)]
-pub(crate) struct WorkerCacheHitEstimate {
-    pub effective_overlap_blocks: f64,
-    pub cached_tokens: usize,
-}
-
-impl WorkerCacheHitEstimate {
-    pub fn rounded_overlap_blocks(self) -> u32 {
-        self.effective_overlap_blocks.round() as u32
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-struct CacheHitEstimates {
-    effective_overlap_blocks: HashMap<WorkerWithDpRank, f64>,
-    cached_tokens: HashMap<WorkerWithDpRank, usize>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct WorkerOverlapScore {
-    pub worker_id: WorkerId,
-    pub dp_rank: DpRank,
-    pub device_blocks: usize,
-    pub host_pinned_blocks: usize,
-    pub disk_blocks: usize,
-    pub host_pinned_extension_blocks: usize,
-    pub disk_extension_blocks: usize,
-    pub shared_beyond_device_blocks: Option<u32>,
-    pub router_credit_blocks: f64,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct SharedCacheOverlapScore {
-    pub enabled: bool,
-    pub total_hit_blocks: u32,
-    pub ranges: Vec<(u32, u32)>,
-    pub error: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct OverlapScoresResponse {
-    pub block_size: u32,
-    pub num_blocks: usize,
-    pub workers: Vec<WorkerOverlapScore>,
-    pub shared_cache: SharedCacheOverlapScore,
-}
-
-fn cache_hit_weight_for_tier(
-    kv_router_config: &KvRouterConfig,
-    storage_tier: dynamo_kv_router::protocols::StorageTier,
-) -> f64 {
-    match storage_tier {
-        dynamo_kv_router::protocols::StorageTier::Device => 1.0,
-        dynamo_kv_router::protocols::StorageTier::HostPinned => {
-            kv_router_config.host_cache_hit_weight
-        }
-        dynamo_kv_router::protocols::StorageTier::Disk
-        | dynamo_kv_router::protocols::StorageTier::External => {
-            kv_router_config.disk_cache_hit_weight
-        }
-    }
-}
-
-fn cached_tokens_from_effective_overlap(block_size: u32, effective_overlap_blocks: f64) -> usize {
-    (effective_overlap_blocks * block_size as f64)
-        .round()
-        .max(0.0) as usize
-}
-
-fn cache_hit_estimates_from_tiered_matches(
-    kv_router_config: &KvRouterConfig,
-    block_size: u32,
-    tiered_matches: &indexer::TieredMatchDetails,
-) -> CacheHitEstimates {
-    let mut effective_overlap_blocks = HashMap::new();
-
-    for (worker, overlap) in &tiered_matches.device.overlap_scores.scores {
-        effective_overlap_blocks.insert(*worker, *overlap as f64);
-    }
-
-    for (storage_tier, tier_matches) in &tiered_matches.lower_tier {
-        let weight = cache_hit_weight_for_tier(kv_router_config, *storage_tier);
-        if weight == 0.0 {
-            continue;
-        }
-
-        for (worker, hits) in &tier_matches.hits {
-            if *hits == 0 {
-                continue;
-            }
-            *effective_overlap_blocks.entry(*worker).or_insert(0.0) += *hits as f64 * weight;
-        }
-    }
-
-    let cached_tokens = effective_overlap_blocks
-        .iter()
-        .map(|(worker, overlap)| {
-            (
-                *worker,
-                cached_tokens_from_effective_overlap(block_size, *overlap),
-            )
-        })
-        .collect();
-
-    CacheHitEstimates {
-        effective_overlap_blocks,
-        cached_tokens,
-    }
-}
-
-fn cache_hit_for_worker(
-    cache_hit_estimates: &CacheHitEstimates,
-    worker: WorkerWithDpRank,
-) -> WorkerCacheHitEstimate {
-    WorkerCacheHitEstimate {
-        effective_overlap_blocks: cache_hit_estimates
-            .effective_overlap_blocks
-            .get(&worker)
-            .copied()
-            .unwrap_or(0.0),
-        cached_tokens: cache_hit_estimates
-            .cached_tokens
-            .get(&worker)
-            .copied()
-            .unwrap_or(0),
-    }
-}
-
-fn tier_overlap_blocks_from_tiered_matches(
-    tiered_matches: &indexer::TieredMatchDetails,
-) -> TierOverlapBlocks {
-    let mut tier_overlap_blocks = TierOverlapBlocks::default();
-
-    tier_overlap_blocks.device.extend(
-        tiered_matches
-            .device
-            .overlap_scores
-            .scores
-            .iter()
-            .map(|(worker, hits)| (*worker, *hits as usize)),
-    );
-
-    if let Some(host_matches) = tiered_matches
-        .lower_tier
-        .get(&dynamo_kv_router::protocols::StorageTier::HostPinned)
-    {
-        tier_overlap_blocks.host_pinned.extend(
-            host_matches
-                .hits
-                .iter()
-                .map(|(worker, hits)| (*worker, *hits)),
-        );
-    }
-
-    // Disk and External share the same weighting (see `storage_tier_weight`),
-    // so accumulate both into the disk bucket.
-    for tier in [
-        dynamo_kv_router::protocols::StorageTier::Disk,
-        dynamo_kv_router::protocols::StorageTier::External,
-    ] {
-        if let Some(matches) = tiered_matches.lower_tier.get(&tier) {
-            for (worker, hits) in &matches.hits {
-                *tier_overlap_blocks.disk.entry(*worker).or_default() += *hits;
-            }
-        }
-    }
-
-    tier_overlap_blocks
-}
-
-fn shared_cache_overlap_score(
-    enabled: bool,
-    hits: Option<&dynamo_kv_router::protocols::SharedCacheHits>,
-    error: Option<String>,
-) -> SharedCacheOverlapScore {
-    let Some(hits) = hits else {
-        return SharedCacheOverlapScore {
-            enabled,
-            total_hit_blocks: 0,
-            ranges: Vec::new(),
-            error,
-        };
-    };
-
-    SharedCacheOverlapScore {
-        enabled,
-        total_hit_blocks: hits.total_hits,
-        ranges: hits
-            .ranges
-            .iter()
-            .map(|range| (range.start, range.end))
-            .collect(),
-        error,
-    }
-}
-
-/// Re-queries the indexer to refresh stale overlap scores for requests that have been
-/// waiting in the scheduler queue. Wired into the scheduler's
-/// [`OverlapScoresRefresh`](dynamo_kv_router::scheduling::OverlapScoresRefresh) slot at
-/// router startup.
-///
-/// `Remote` and `None` indexer variants intentionally don't get a refresher: remote
-/// indexer queries are too expensive to repeat per request, and `None` has nothing to query.
-struct KvRouterOverlapRefresher {
-    indexer: Indexer,
-    kv_router_config: KvRouterConfig,
-    block_size: u32,
-}
-
-impl KvRouterOverlapRefresher {
-    fn for_indexer(
-        indexer: Indexer,
-        kv_router_config: KvRouterConfig,
-        block_size: u32,
-    ) -> Option<Self> {
-        match &indexer {
-            Indexer::KvIndexer { .. } | Indexer::Concurrent { .. } => Some(Self {
-                indexer,
-                kv_router_config,
-                block_size,
-            }),
-            Indexer::Remote { .. } | Indexer::None => None,
-        }
-    }
-}
-
-#[async_trait]
-impl OverlapScoresRefresh for KvRouterOverlapRefresher {
-    async fn refresh(&self, block_hashes: &[LocalBlockHash]) -> Option<RefreshedOverlap> {
-        let tiered = match self.indexer.find_matches_by_tier_ref(block_hashes).await {
-            Ok(t) => t,
-            Err(e) => {
-                tracing::warn!(error = ?e, "overlap refresh: find_matches_by_tier failed");
-                return None;
-            }
-        };
-        let tier_overlap_blocks = tier_overlap_blocks_from_tiered_matches(&tiered);
-        let estimates = cache_hit_estimates_from_tiered_matches(
-            &self.kv_router_config,
-            self.block_size,
-            &tiered,
-        );
-        Some(RefreshedOverlap {
-            tier_overlap_blocks,
-            effective_overlap_blocks: estimates.effective_overlap_blocks,
-            effective_cached_tokens: estimates.cached_tokens,
-        })
-    }
-}
 
 fn map_scheduler_error(error: scheduling::KvSchedulerError) -> anyhow::Error {
     if !error.is_overload() {
@@ -694,104 +432,21 @@ where
         let supports_overlap_refresh = self.scheduler.supports_overlap_refresh();
         let retain_block_hashes = supports_overlap_refresh || return_routing_hashes;
 
-        let (
+        let TieredLookupResult {
             tiered_matches,
             shared_cache_hits,
             indexer_duration,
             shared_cache_duration,
             retained_block_hashes,
-        ) = if retain_block_hashes {
-            // Query indexer (tiered) and shared cache in parallel when shared cache is configured.
-            // Time each independently so metrics can separate indexer vs shared cache latency.
-            let (tiered, shared_hits, idx_dur, sc_dur) =
-                if let Some(ref shared_cache) = self.shared_cache {
-                    let indexer_fut = self
-                        .indexer
-                        .find_matches_by_tier_ref(&block_hashes)
-                        .instrument(tracing::info_span!("kv_router.find_matches"));
-                    let shared_fut = shared_cache
-                        .check_blocks(tokens, self.block_size)
-                        .instrument(tracing::info_span!("kv_router.shared_cache_check"));
-
-                    let indexer_timed = async {
-                        let t = Instant::now();
-                        let r = indexer_fut.await;
-                        (r, t.elapsed())
-                    };
-                    let shared_timed = async {
-                        let t = Instant::now();
-                        let r = shared_fut.await;
-                        (r, t.elapsed())
-                    };
-
-                    let ((indexer_result, idx_dur), (shared_result, sc_dur)) =
-                        tokio::join!(indexer_timed, shared_timed);
-                    let tiered = indexer_result?;
-                    // Shared cache failure is non-fatal: log warning and fall back to empty hits.
-                    let hits = match shared_result {
-                        Ok(hits) => Some(hits),
-                        Err(e) => {
-                            tracing::warn!(error = %e, "Shared cache query failed, ignoring");
-                            if let Some(m) = metrics::RoutingOverheadMetrics::get() {
-                                m.inc_shared_cache_errors();
-                            }
-                            None
-                        }
-                    };
-                    (tiered, hits, idx_dur, Some(sc_dur))
-                } else {
-                    let t = Instant::now();
-                    let tiered = self
-                        .indexer
-                        .find_matches_by_tier_ref(&block_hashes)
-                        .instrument(tracing::info_span!("kv_router.find_matches"))
-                        .await?;
-                    (tiered, None, t.elapsed(), None)
-                };
-            (tiered, shared_hits, idx_dur, sc_dur, Some(block_hashes))
-        } else if let Some(ref shared_cache) = self.shared_cache {
-            let indexer_fut = self
-                .indexer
-                .find_matches_by_tier(block_hashes)
-                .instrument(tracing::info_span!("kv_router.find_matches"));
-            let shared_fut = shared_cache
-                .check_blocks(tokens, self.block_size)
-                .instrument(tracing::info_span!("kv_router.shared_cache_check"));
-
-            let indexer_timed = async {
-                let t = Instant::now();
-                let r = indexer_fut.await;
-                (r, t.elapsed())
-            };
-            let shared_timed = async {
-                let t = Instant::now();
-                let r = shared_fut.await;
-                (r, t.elapsed())
-            };
-
-            let ((indexer_result, idx_dur), (shared_result, sc_dur)) =
-                tokio::join!(indexer_timed, shared_timed);
-            let tiered = indexer_result?;
-            let hits = match shared_result {
-                Ok(hits) => Some(hits),
-                Err(e) => {
-                    tracing::warn!(error = %e, "Shared cache query failed, ignoring");
-                    if let Some(m) = metrics::RoutingOverheadMetrics::get() {
-                        m.inc_shared_cache_errors();
-                    }
-                    None
-                }
-            };
-            (tiered, hits, idx_dur, Some(sc_dur), None)
-        } else {
-            let t = Instant::now();
-            let tiered = self
-                .indexer
-                .find_matches_by_tier(block_hashes)
-                .instrument(tracing::info_span!("kv_router.find_matches"))
-                .await?;
-            (tiered, None, t.elapsed(), None, None)
-        };
+        } = query_tiered_matches(
+            &self.indexer,
+            self.shared_cache.as_deref(),
+            tokens,
+            self.block_size,
+            block_hashes,
+            retain_block_hashes,
+        )
+        .await?;
 
         let (block_hashes_for_refresh, routing_block_hashes) = retained_block_hashes
             .map(|block_hashes| {
